# fix: Improve timing tolerance in performance test